lots of documentation cleanup

+++ sp-form/fields/sp-dropzone-field/ViewModel.js
@@ -5,6 +5,12 @@ import DefineList from 'can-define/list/list';
 import DefineMap from 'can-define/map/map';
 
 /**
+ * The dropzone library object
+ * See {@link http://www.dropzonejs.com/}
+ * @typedef {Object} Dropzone
+ */
+
+/**
  * File type for dropzone field and file list 
  * @class File
  * @memberof sp-dropzone-field
@@ -42,10 +48,37 @@ const FileMapList = DefineList.extend('FileList', {
  */
 export default Field.extend('DropZoneField', {
     /** @lends sp-dropzone-field.ViewModel.prototype */
+    
+    /**
+     * @type {Array<FileMap>}
+     * @memberof sp-dropzone-field.ViewModel.prototype
+     * @description The current array of files
+     */
     value: {Default: FileMapList, Type: FileMapList},
+    /**
+     * @type {Dropzone}
+     * @memberof sp-dropzone-field.ViewModel.prototype
+     * @description The dropzone library object. 
+     * @see {@link http://www.dropzonejs.com/|Dropzone}
+     */
     dropzone: '*',
+    /**
+     * @type {Array<FileMap>}
+     * @memberof sp-dropzone-field.ViewModel.prototype
+     * @description The current array of files
+     */
     url: {type: 'string', default: '/api/uploads'},
+    /**
+     * @type {Array<FileMap>}
+     * @memberof sp-dropzone-field.ViewModel.prototype
+     * @description The current array of files
+     */
     paramName: {type: 'string', default: 'uri'},
+    /**
+     * @type {Array<FileMap>}
+     * @memberof sp-dropzone-field.ViewModel.prototype
+     * @description The current array of files
+     */
     headers: {
         type: '*', 
         default () { 
@@ -55,6 +88,20 @@ export default Field.extend('DropZoneField', {
             };
         }
     },
+    /**
+     * @type {Number}
+     * @memberof sp-dropzone-field.ViewModel.prototype
+     * @description The timout to automatically remove files from the dropzone container
+     * after successful uploads. Set to 0 to never remove
+     */
+    timeout: {
+        type: 'number',
+        default: 2000
+    },
+    /**
+     * Initializes the dropzone
+     * @param {Element} element the element to initialize dropzone on
+     */
     createDropzone (element) {
         this.dropzone = new Dropzone(element, {
             url: this.url,
@@ -66,6 +113,12 @@ export default Field.extend('DropZoneField', {
         this.dropzone.on('success', this.uploadSuccess.bind(this));
         this.dropzone.on('complete', this.clearUploadedFile.bind(this));
     },
+    /**
+     * 
+     * When the upload succeeds, the 'fieldchange' event is dispached
+     * @param {FileMap} file The uploaded file
+     * @param {*} result The result json from the rest service
+     */
     uploadSuccess (file, result) {
         if (typeof result === 'string') {
             try {
@@ -80,11 +133,23 @@ export default Field.extend('DropZoneField', {
         this.value.push(result);
         this.dispatch('fieldchange', [this]);
     },
+    /**
+     * Removes the uploaded file after a set time
+     * @param {FileMap} file the file to remove
+     */
     clearUploadedFile (file) {
-        setTimeout(() => {
-            this.dropzone.removeFile(file);
-        }, 2000);
+        if (this.timeout) {
+            setTimeout(() => {
+                this.dropzone.removeFile(file);
+            }, this.timeout); 
+        }
     },
+    /**
+     * 
+     * Sends a DELETE request to the api endpoint to remove the file using Axios
+     * @param {FileMap} file The file to delete
+     * @returns {Promise<AxiosResult>}
+     */
     delete (file) {
         file.isDeleting = true;
         
@@ -97,22 +162,21 @@ export default Field.extend('DropZoneField', {
 
         return axios.delete(this.url + '/' + file.id, {
             headers: this.headers
-        }).then(() => {
+        }).then((result) => {
             // remove the file from our store on successful delete
             this.value.splice(this.value.indexOf(file), 1);
             this.dispatch('fieldchange', [this]);
+            return result;
         }).catch((e) => {
             // if status is a 404, the file wasn't found anyways, so we 
             // should remove it from the store and perhaps log a warning
             if (e.request && e.request.status === 404) {
                 this.value.splice(this.value.indexOf(file), 1);
             }
-            // eslint-disable-next-line
-            //!steal-remove-start
+            // !steal-remove-start
             // eslint-disable-next-line
             console.warn('error occurred when deleting file', file, e);
-            // eslint-disable-next-line
-            //!steal-remove-end
+            // !steal-remove-end
         });
     }
 });
